Delete old tensor worker bootstrap
This is only used in the old workers. Trimming down tensor_engine_worker so we can remove torch dependencies more easily.

Differential Revision: [D87574093](https://our.internmc.facebook.com/intern/diff/D87574093/)

[ghstack-poisoned]
